refactor: uniformize file naming convention to kebab-case
- Renamed all snake_case files to kebab-case for consistency
- Updated all import paths throughout the codebase
- Maintained functionality while improving code organization
- Fixed critical code quality issue identified in audit
