+++ src/components/common/bottom-blur.tsx
