+++ src/components/effects/profile-background.tsx
