+++ src/components/effects/smile-grid.tsx
