+++ src/components/layout/background-gradient.tsx
