+++ src/components/sections/contact.tsx
@@ -1,5 +1,5 @@
-import FadeInUpOnScroll from "@/components/common/animations/fade_in_up_on_scroll";
-import LetterRippleEffect from "@/components/effects/letter_ripple";
+import FadeInUpOnScroll from "@/components/common/animations/fade-in-up-on-scroll";
+import LetterRippleEffect from "@/components/effects/letter-ripple";
 import Hand from "@/assets/images/ui/hand.svg?react";
 import { motion } from "framer-motion";
